fix(cli): stabilize mock MCP test server
