+++ rust/crates/astra-cli/Cargo.toml
@@ -7,6 +7,10 @@ edition.workspace = true
 name = "astra"
 path = "src/main.rs"
 
+[[bin]]
+name = "mock_mcp_server"
+path = "src/bin/mock_mcp_server.rs"
+
 [features]
 default = ["harness"]
 harness = ["astra-runtime/harness", "dep:astra-harness"]
@@ -71,8 +75,10 @@ tree-sitter-go = "0.23"
 tree-sitter-java = "0.23"
 tree-sitter-cpp = "0.23"
 tree-sitter-ruby = "0.23"
-# MCP client for external tool integration
-rmcp = { version = "1.3", default-features = false, features = ["client", "transport-child-process", "transport-streamable-http-client-reqwest", "transport-async-rw"] }
+# MCP client for external tool integration, plus the lightweight mock stdio
+# server binary used by tests.
+rmcp = { version = "1.3", default-features = false, features = ["client", "server", "macros", "schemars", "transport-child-process", "transport-streamable-http-client-reqwest", "transport-async-rw", "transport-io"] }
+schemars = "1"
 tokio-tungstenite = { version = "0.26", features = ["rustls-tls-webpki-roots"] }
 nix.workspace = true
 pulldown-cmark = "0.12"
@@ -97,6 +103,3 @@ serial_test = "3"
 temp-env = "0.3"
 wiremock = "0.6"
 insta.workspace = true
-# rmcp server features for mock MCP server example (integration tests)
-rmcp = { version = "1.3", default-features = false, features = ["server", "macros", "schemars", "transport-io"] }
-schemars = "1"
