+++ rust/crates/astra-cli/src/mcp_client.rs
@@ -2718,69 +2718,65 @@ pub(crate) fn ensure_mock_mcp_server_binary() -> std::path::PathBuf {
                 .join("../..")
                 .join("target")
                 .join(profile)
-                .join("examples")
                 .join(&binary_name);
             let lock_path = binary.with_file_name(format!("{binary_name}.build.lock"));
 
-            if binary.exists() {
-                return binary;
-            }
-
-            std::fs::create_dir_all(
-                binary
-                    .parent()
-                    .expect("mock_mcp_server example binary always has a parent directory"),
-            )
-            .unwrap_or_else(|e| {
-                panic!(
-                    "failed to create mock_mcp_server example directory {:?}: {e}",
-                    binary.parent()
+            if !binary.exists() {
+                std::fs::create_dir_all(
+                    binary
+                        .parent()
+                        .expect("mock_mcp_server binary always has a parent directory"),
                 )
-            });
-
-            let mut lock_file = OpenOptions::new()
-                .create(true)
-                .truncate(false)
-                .read(true)
-                .write(true)
-                .open(&lock_path)
                 .unwrap_or_else(|e| {
                     panic!(
-                        "failed to open mock_mcp_server build lock {:?}: {e}",
+                        "failed to create mock_mcp_server directory {:?}: {e}",
+                        binary.parent()
+                    )
+                });
+
+                let mut lock_file = OpenOptions::new()
+                    .create(true)
+                    .truncate(false)
+                    .read(true)
+                    .write(true)
+                    .open(&lock_path)
+                    .unwrap_or_else(|e| {
+                        panic!(
+                            "failed to open mock_mcp_server build lock {:?}: {e}",
+                            lock_path
+                        )
+                    });
+                lock_file.lock_exclusive().unwrap_or_else(|e| {
+                    panic!(
+                        "failed to acquire mock_mcp_server build lock {:?}: {e}",
                         lock_path
                     )
                 });
-            lock_file.lock_exclusive().unwrap_or_else(|e| {
-                panic!(
-                    "failed to acquire mock_mcp_server build lock {:?}: {e}",
-                    lock_path
-                )
-            });
 
-            if binary.exists() {
-                return binary;
+                if !binary.exists() {
+                    let _ = lock_file.set_len(0);
+                    let _ = writeln!(
+                        lock_file,
+                        "building mock_mcp_server for pid {}",
+                        std::process::id()
+                    );
+
+                    let status = std::process::Command::new("cargo")
+                        .args(["build", "-p", "astra-cli", "--bin", "mock_mcp_server"])
+                        .current_dir(manifest_dir.join("../.."))
+                        .status()
+                        .unwrap_or_else(|e| panic!("failed to build mock_mcp_server bin: {e}"));
+
+                    assert!(
+                        status.success(),
+                        "cargo build -p astra-cli --bin mock_mcp_server failed with status {status}"
+                    );
+                }
             }
 
-            let _ = lock_file.set_len(0);
-            let _ = writeln!(
-                lock_file,
-                "building mock_mcp_server for pid {}",
-                std::process::id()
-            );
-
-            let status = std::process::Command::new("cargo")
-                .args(["build", "-p", "astra-cli", "--example", "mock_mcp_server"])
-                .current_dir(manifest_dir.join("../.."))
-                .status()
-                .unwrap_or_else(|e| panic!("failed to build mock_mcp_server example: {e}"));
-
-            assert!(
-                status.success(),
-                "cargo build -p astra-cli --example mock_mcp_server failed with status {status}"
-            );
             assert!(
                 binary.exists(),
-                "mock_mcp_server binary missing after build at {:?}",
+                "mock_mcp_server binary missing at {:?} after prebuild or fallback build",
                 binary
             );
